+++ packages/ui-components/src/components/Maps/USNationalMap/CanvasMap.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useRef } from "react";
 import { ExtendedFeature, GeoProjection, geoPath as d3GeoPath } from "d3-geo";
-import { StyledCanvas } from "./NewUSNationalMap.style";
+import { StyledCanvas } from "./USNationalMap.style";
 import { stateBorders } from "../../../common/geo-shapes";
 
 const borderColor = "black";
