+++ packages/ui-components/src/components/Maps/USNationalMap/StatesMap.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { GeoPath } from "d3-geo";
 import Tooltip from "@mui/material/Tooltip";
-import { StateOverlay, StateShape } from "./NewUSNationalMap.style";
+import { StateOverlay, StateShape } from "./USNationalMap.style";
 import { statesGeographies } from "../../../common/geo-shapes";
 import { states } from "@actnowcoalition/regions";
 
@@ -9,9 +9,9 @@ const StatesMap: React.FC<{
   width: number;
   height: number;
   geoPath: GeoPath;
-  renderTooltip: (fips: string) => React.ReactElement | string;
+  renderTooltip: (regionId: string) => React.ReactElement | string;
   showCounties: boolean;
-  getFillColor: (fips: string) => string;
+  getFillColor: (regionId: string) => string;
 }> = ({
   width,
   height,
