+++ packages/ui-components/src/components/Maps/USNationalMap/USNationalMap.tsx
@@ -1,77 +1,66 @@
-import React, { useMemo } from "react";
-import { states } from "@actnowcoalition/regions";
-import Tooltip from "@mui/material/Tooltip";
-import { geoAlbersUsaTerritories } from "geo-albers-usa-territories";
-import { ComposableMap, Geographies } from "react-simple-maps";
-import { statesGeographies } from "../../../common/geo-shapes";
+import React from "react";
+import { MapContainer, PositionAbsolute } from "./USNationalMap.style";
+import {
+  defaultHeight,
+  defaultScale,
+  defaultWidth,
+} from "../../../common/geo-shapes";
+import { getCountyGeoId } from "../../../common/geo-shapes";
+import { geoPath as d3GeoPath, geoAlbersUsa } from "d3-geo";
+import StatesMap from "./StatesMap";
 import CountiesMap from "./CountiesMap";
-import { StateGeoPath } from "./USNationalMap.style";
 
 export interface USNationalMapProps {
-  renderTooltip: (fips: string) => React.ReactElement | string;
-  getFillColor?: (fips: string) => string;
+  width: number;
+  renderTooltip: (regionId: string) => React.ReactElement | string;
+  getFillColor?: (regionId: string) => string;
   showCounties?: boolean;
 }
 
-const projection = geoAlbersUsaTerritories().scale(1070).translate([400, 250]);
-
-const statesFipsCodes = states.all.map((state) => state.regionId);
-
-const USNationalMapFC: React.FC<USNationalMapProps> = ({
-  getFillColor,
+const USNationalMap: React.FC<USNationalMapProps> = ({
+  width = defaultWidth,
   renderTooltip,
+  getFillColor = () => "white",
   showCounties = false,
 }) => {
-  const regionGeographies = useMemo(() => {
-    return (
-      <>
-        <Geographies geography={statesGeographies}>
-          {({ geographies }) =>
-            geographies
-              .filter((geo) => statesFipsCodes.includes(geo.id))
-              .map((geo) => {
-                const fipsCode = geo.id;
-                const stateRegion = states.findByRegionIdStrict(fipsCode);
-                const stateName = stateRegion.fullName;
+  const height = defaultHeight * (width / defaultWidth);
+  const scale = (defaultScale * width) / defaultWidth;
 
-                const geography = (
-                  <StateGeoPath
-                    key={fipsCode}
-                    geography={geo}
-                    /* If we are rendering counties, we do not color the state shapes */
-                    fill={
-                      !getFillColor || showCounties
-                        ? "white"
-                        : getFillColor(geo.id)
-                    }
-                    stroke="black"
-                    strokeOpacity={1}
-                    strokeWidth={1}
-                    role="img"
-                    tabIndex={-1} // TODO: revisit making the map more accessible. For now removing from tab index to avoid a focus trap
-                    aria-label={stateName}
-                  />
-                );
+  const projection = geoAlbersUsa()
+    .scale(scale)
+    .translate([width / 2, height / 2]);
 
-                return (
-                  <Tooltip key={fipsCode} title={renderTooltip(fipsCode)}>
-                    {geography}
-                  </Tooltip>
-                );
-              })
-          }
-        </Geographies>
-        {showCounties && <CountiesMap getFillColor={getFillColor} />}
-      </>
-    );
-  }, [getFillColor, renderTooltip, showCounties]);
+  const canvasProjection = geoAlbersUsa()
+    .scale(2 * scale)
+    .translate([width, height]);
+
+  const geoPath = d3GeoPath().projection(projection);
 
   return (
-    <ComposableMap projection={projection} height={500}>
-      {regionGeographies}
-    </ComposableMap>
+    <MapContainer>
+      {showCounties && (
+        <PositionAbsolute>
+          <CountiesMap
+            width={width}
+            height={height}
+            getFillColor={getFillColor}
+            geoProjection={canvasProjection}
+            getGeoId={getCountyGeoId}
+          />
+        </PositionAbsolute>
+      )}
+      <PositionAbsolute>
+        <StatesMap
+          width={width}
+          height={height}
+          geoPath={geoPath}
+          renderTooltip={renderTooltip}
+          showCounties={showCounties}
+          getFillColor={getFillColor}
+        />
+      </PositionAbsolute>
+    </MapContainer>
   );
 };
 
-const USNationalMap = React.memo(USNationalMapFC);
 export default USNationalMap;
