+++ packages/ui-components/src/components/Maps/index.ts
@@ -1,1 +1,1 @@
-export { default as USNationalMap } from "./USNationalMap/USNationalMap";
+export { default as USNationalMap } from "./USNationalMap";
